requirements.txt
discord.py[voice]==1.7.1